requirements.txt
# paddleocr==2.7.0.3
pdfplumber==0.10.3
python-dotenv==1.0.0
pdf2docx==0.5.6
mammoth==1.6.0
flask==3.0.0
pandas==2.1.4
python-dateutil==2.8.2
openai==1.6.1
requests==2.31.0